Bitmap layer renamed due to overlapping
