+++ src/plugins/plugin-mapbox-gl/custom-layers/decode-layer/decode-layer-fragment.js
